Modify: config.js 포함한 코드 리팩토링 - config.js 파일 생성하여 API 일괄적으로 관리 가능 - Login.scss에 있는 중복 내용 삭제 - handleDelete 함수 로직 변경 - 컴포넌트 중복 제거
