+++ src/components/Login/Login.js
@@ -1,5 +1,8 @@
 import React, { Component } from 'react';
+import Email from './Email';
+import Password from './Password';
 import { withRouter } from 'react-router-dom';
+import { GET_PRODUCT_API } from '../../config';
 import './Login.scss';
 
 class Login extends Component {
@@ -12,7 +15,7 @@ class Login extends Component {
   }
 
   handleLogin = () => {
-    fetch('http://172.24.173.9:8000/users/login', {
+    fetch(`${GET_PRODUCT_API}/login`, {
       method: 'POST',
       body: JSON.stringify({
         email: this.state.email,
@@ -36,7 +39,7 @@ class Login extends Component {
   };
 
   handleDelete = e => {
-    const name = e.target.parentNode.childNodes[0].name;
+    const name = e.target.name;
     this.setState({ [name]: '' });
   };
 
@@ -44,93 +47,26 @@ class Login extends Component {
     const { email, pw } = this.state;
     const { handleInput, handleDelete, handleLogin } = this;
 
-    const checkEmail = email => {
-      const regExp = /^[a-zA-Z\d+-.]+@[a-zA-Z\d+-.]+\.[a-zA-Z]{2,3}$/;
-      return regExp.test(email);
-    };
-
-    const checkPw = pw => {
-      const regExp =
-        /^(?=.*[A-Za-z])(?=.*\d)(?=.*[!@#$%^&*-_])[A-Za-z\d!@#$%^&*-_]{10,}$/;
-      return regExp.test(pw);
-    };
-
     return (
       <div className="login">
         <div className="login-container">
           <div className="login-logo">
-            <img
-              src="https://upload.wikimedia.org/wikipedia/commons/b/b8/%EC%99%93%EC%B1%A0_%EB%A1%9C%EA%B3%A0_2021.png"
-              alt="logo"
-            />
+            <img src="images/AtchaPedia_Logo_Big.png" alt="logo" />
           </div>
           <div className="login-title">
             <span>로그인</span>
           </div>
           <div className="login-form">
-            <div
-              className={`email-container ${
-                !email || checkEmail(email) ? '' : 'invalid'
-              }`}
-            >
-              <input
-                id="email"
-                name="email"
-                className="email"
-                placeholder="이메일"
-                type="text"
-                value={email}
-                onChange={handleInput}
-              />
-              <i
-                className={`fas fa-times-circle ${!email ? 'none' : ''}`}
-                onClick={handleDelete}
-              ></i>
-              <i
-                className={`far fa-check-circle ${!email ? 'none' : ''}`}
-                style={{
-                  color: checkEmail(email) ? '#4ad3b1' : '#a0a0a0',
-                }}
-              ></i>
-            </div>
-            <div className="email-validation">
-              <p
-                className={`email-caution ${
-                  !email || checkEmail(email) ? 'none' : ''
-                }`}
-              >
-                정확하지 않은 이메일입니다.
-              </p>
-            </div>
-            <div
-              className={`pw-container ${!pw || checkPw(pw) ? '' : 'invalid'}`}
-            >
-              <input
-                id="pw"
-                name="pw"
-                className="password"
-                placeholder="비밀번호"
-                type="password"
-                value={pw}
-                onChange={handleInput}
-              />
-              <i
-                className={`fas fa-times-circle ${!pw ? 'none' : ''}`}
-                onClick={handleDelete}
-              ></i>
-              <i
-                className={`far fa-check-circle ${!pw ? 'none' : ''}`}
-                style={{
-                  color: checkPw(pw) ? '#4ad3b1' : '#a0a0a0',
-                }}
-              ></i>
-            </div>
-            <div className="pw-validation">
-              <p className={`pw-caution ${!pw || checkPw(pw) ? 'none' : ''}`}>
-                비밀번호는 영문, 숫자, 특수문자 모두 포함하여
-                <br /> 최소 10자리 이상이여야 합니다.
-              </p>
-            </div>
+            <Email
+              email={email}
+              handleInput={handleInput}
+              handleDelete={handleDelete}
+            />
+            <Password
+              pw={pw}
+              handleInput={handleInput}
+              handleDelete={handleDelete}
+            />
             <button className="btn-login" onClick={handleLogin}>
               로그인
             </button>
