+++ src/components/Login/Login.scss
@@ -1,21 +1,4 @@
-// font-size
-$font-size-huge: 17px;
-$font-size-large: 16px;
-$font-size-medium: 15px;
-$font-size-small: 14px;
-$font-size-micro: 13px;
-
-// color
-$color-background: #f8f8f8;
-$color-red: #f50700;
-$color-pink: #f52f6e;
-$color-green: #4ad3b1;
-$color-light-pink: #fef0f0;
-$color-black: #000000;
-$color-light-black: #292a32;
-$color-white: #ffffff;
-$color-gray: #a0a0a0;
-$color-light-gray: #ebebeb;
+@import '../../styles/variables.scss';
 
 .login {
   width: 375px;
