+++ src/config.js
@@ -1,0 +1,3 @@
+// Login
+const BASE_URL = 'http://10.58.6.81:8000';
+export const GET_PRODUCT_API = `${BASE_URL}/users`;
